src/requirements.txt
apturl==0.5.2
asgiref==3.5.2
bcrypt==3.1.7
blinker==1.4
blis==0.2.4
Brlapi==0.7.0
certifi==2019.11.28
cffi==1.15.1
chardet==3.0.4
ChatterBot==1.0.5
click==8.1.3
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
cymem==2.0.6
dbus-python==1.2.16
defer==1.0.6
distro==1.4.0
distro-info===0.23ubuntu1
duplicity==0.8.12.0
entrypoints==0.3
fasteners==0.14.1
Flask==2.2.2
Flask-Cors==3.0.10
future==0.18.2
httplib2==0.14.0
idna==2.8
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.1.0
kazam==1.4.5
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==2.1.1
mathparse==0.1.2
monotonic==1.5
murmurhash==1.0.8
mysql-connector==2.2.9
mysql-connector-python==8.0.28
netifaces==0.10.4
nltk==3.7
numpy==1.22.3
oauthlib==3.1.0
paramiko==2.6.0
pexpect==4.6.0
Pint==0.19.2
plac==0.9.6
preshed==2.0.1
protobuf==3.20.0
PyAudio==0.2.11
pycairo==1.16.2
pycparser==2.21
pycups==1.9.73
pydub==0.25.1
pygame==2.1.2
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
pymongo==3.12.3
PyNaCl==1.3.0
pyRFC3339==1.1
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.7.3
python-debian==0.1.36+ubuntu1.1
pytz==2019.3
pyxdg==0.26
PyYAML==5.1.2
regex==2022.8.17
requests==2.22.0
requests-unixsocket==0.2.0
scipy==1.8.0
SecretStorage==2.3.1
simplejson==3.16.0
six==1.14.0
SoundFile==0.10.3.post1
sox==1.4.1
spacy==2.1.9
SpeechRecognition==3.8.1
SQLAlchemy==1.2.19
srsly==1.0.5
system-service==0.3
systemd-python==234
thinc==7.0.8
torch==1.12.1
tqdm==4.64.0
typing-extensions==4.3.0
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
wadllib==1.3.3
wasabi==0.10.1
websockets==10.2
Werkzeug==2.2.2
xkit==0.0.0
zipp==3.8.1
